changed type

+++ cv2utils/video.py
@@ -10,7 +10,7 @@ class cv2_video:
         self.height = self.cap.get(cv2.CAP_PROP_FRAME_HEIGHT)
         self.width = self.cap.get(cv2.CAP_PROP_FRAME_WIDTH)
 
-        if type(self.fps) == float and self.fps > 0 and type(self.total_frame) == int:
+        if type(self.fps) == float and self.fps > 0 and type(self.total_frame) == float:
             self.duration = int(self.total_frame / self.fps)
         else:
             warnings.warn("can't define video duration")
